Isolated fetching graduation progress information into utils file
